WIP Sort by rooms underpopulated for spawning Moved RoleBodyParts to rolesManager.ts as it's more explicit
